Split NativeMemoryDiagnoser into two classes
