+++ src/BenchmarkDotNet.Diagnostics.Windows/Tracing/NativeMemoryLogParser.cs
@@ -1,0 +1,198 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Loggers;
+using BenchmarkDotNet.Reports;
+using BenchmarkDotNet.Running;
+using Microsoft.Diagnostics.Tracing.Etlx;
+using Microsoft.Diagnostics.Tracing.Parsers.Kernel;
+using Address = System.UInt64;
+
+namespace BenchmarkDotNet.Diagnostics.Windows.Tracing
+{
+    public class NativeMemoryLogParser
+    {
+        private static readonly string LogSeparator = new string('-', 20);
+        private readonly string etlFilePath;
+        private readonly BenchmarkCase benchmarkCase;
+
+        private readonly ILogger logger;
+
+        public NativeMemoryLogParser(string etlFilePath, BenchmarkCase benchmarkCase, ILogger logger)
+        {
+            this.etlFilePath = etlFilePath;
+            this.benchmarkCase = benchmarkCase;
+            this.logger = logger;
+        }
+        public IEnumerable<Metric> Parse()
+        {
+            return TraceLogParserHelper.Parse(etlFilePath, ParseTraceEventSource);
+        }
+
+        //Code is inspired by https://github.com/Microsoft/perfview/blob/master/src/PerfView/PerfViewData.cs#L5719-L5944
+        private IEnumerable<Metric> ParseTraceEventSource(TraceLogEventSource traceLogEventSource)
+        {
+            var bdnEventsParser = new EngineEventLogParser(traceLogEventSource);
+
+            var start = false;
+            var isFirstActualStartEnd = false;
+
+            long totalOperation = 0;
+
+            bdnEventsParser.WorkloadActualStart += data =>
+            {
+                if (!isFirstActualStartEnd)
+                {
+                    start = true;
+                }
+
+                totalOperation = data.TotalOperations;
+            };
+            bdnEventsParser.WorkloadActualStop += data =>
+            {
+                start = false;
+                isFirstActualStartEnd = true;
+            };
+
+            var heapParser = new HeapTraceProviderTraceEventParser(traceLogEventSource);
+            // We index by heap address and then within the heap we remember the allocation stack
+            var heaps = new Dictionary<Address, Dictionary<Address, long>>();
+            Dictionary<Address, long> lastHeapAllocs = null;
+
+            Address lastHeapHandle = 0;
+
+            long nativeLeakSize = 0;
+            long totalAllocation = 0;
+
+            heapParser.HeapTraceAlloc += delegate (HeapAllocTraceData data)
+            {
+                if (!start)
+                {
+                    return;
+                }
+
+                var allocs = lastHeapAllocs;
+                if (data.HeapHandle != lastHeapHandle)
+                    allocs = CreateHeapCache(data.HeapHandle, heaps, ref lastHeapAllocs, ref lastHeapHandle);
+
+                allocs[data.AllocAddress] = data.AllocSize;
+
+                checked
+                {
+                    nativeLeakSize += data.AllocSize;
+                    totalAllocation += data.AllocSize;
+                }
+            };
+            heapParser.HeapTraceFree += delegate (HeapFreeTraceData data)
+            {
+                if (!start)
+                {
+                    return;
+                }
+                var allocs = lastHeapAllocs;
+                if (data.HeapHandle != lastHeapHandle)
+                {
+                    allocs = CreateHeapCache(data.HeapHandle, heaps, ref lastHeapAllocs, ref lastHeapHandle);
+                }
+
+                if (allocs.TryGetValue(data.FreeAddress, out long alloc))
+                {
+                    nativeLeakSize -= alloc;
+
+                    allocs.Remove(data.FreeAddress);
+                }
+            };
+            heapParser.HeapTraceReAlloc += delegate (HeapReallocTraceData data)
+            {
+                if (!start)
+                {
+                    return;
+                }
+                // Reallocs that actually move stuff will cause a Alloc and delete event
+                // so there is nothing to do for those events.  But when the address is
+                // the same we need to resize
+                if (data.OldAllocAddress != data.NewAllocAddress)
+                {
+                    return;
+                }
+
+                var allocs = lastHeapAllocs;
+                if (data.HeapHandle != lastHeapHandle)
+                {
+                    allocs = CreateHeapCache(data.HeapHandle, heaps, ref lastHeapAllocs, ref lastHeapHandle);
+                }
+
+                // This is a clone of the Free code
+                if (allocs.TryGetValue(data.OldAllocAddress, out long alloc))
+                {
+                    nativeLeakSize -= alloc;
+
+                    allocs.Remove(data.OldAllocAddress);
+                }
+
+                // This is a clone of the Alloc code (sigh don't clone code)
+                allocs[data.NewAllocAddress] = data.NewAllocSize;
+
+                nativeLeakSize += data.NewAllocSize;
+            };
+            heapParser.HeapTraceDestroy += delegate (HeapTraceData data)
+            {
+                if (!start)
+                {
+                    return;
+                }
+
+                // Heap is dieing, kill all objects in it.
+                var allocs = lastHeapAllocs;
+                if (data.HeapHandle != lastHeapHandle)
+                {
+                    allocs = CreateHeapCache(data.HeapHandle, heaps, ref lastHeapAllocs, ref lastHeapHandle);
+                }
+
+                foreach (var alloc in allocs.Values)
+                {
+                    // TODO this is a clone of the free code.
+                    nativeLeakSize -= alloc;
+                }
+            };
+
+            traceLogEventSource.Process();
+
+            logger.WriteLine();
+            logger.WriteLineHeader(LogSeparator);
+            logger.WriteLineInfo($"{benchmarkCase.DisplayInfo}");
+            logger.WriteLineHeader(LogSeparator);
+
+            var totalMemory = totalAllocation / totalOperation;
+            var memoryLeak = nativeLeakSize / totalOperation;
+
+            logger.WriteLine($"Total allocated native memory: {totalMemory:n3}");
+            if (nativeLeakSize != 0)
+            {
+                logger.WriteLine($"Total memory leak: {memoryLeak:n3}");
+            }
+            var heapInfoList = heaps.Select(h => new { Address = h.Key, h.Value.Count, types = h.Value.Values });
+            foreach (var item in heapInfoList.Where(p => p.Count > 0))
+            {
+                logger.WriteLine($"Count of not deallocated object: {item.Count / totalOperation}");
+            }
+
+            yield return new Metric(new AllocatedNativeMemoryDescriptor(), totalMemory);
+            yield return new Metric(new NativeMemoryLeakDescriptor(), memoryLeak);
+        }
+
+        private static Dictionary<Address, long> CreateHeapCache(Address heapHandle, Dictionary<Address, Dictionary<Address, long>> heaps, ref Dictionary<Address, long> lastHeapAllocs, ref Address lastHeapHandle)
+        {
+            Dictionary<Address, long> ret;
+
+            if (!heaps.TryGetValue(heapHandle, out ret))
+            {
+                ret = new Dictionary<Address, long>();
+                heaps.Add(heapHandle, ret);
+            }
+            lastHeapHandle = heapHandle;
+            lastHeapAllocs = ret;
+            return ret;
+        }
+    }
+}
